Move normalize_user_addrs_sorted_impl() fn into normalizer module
This change moves the normalize_user_addrs_sorted_impl() function into
the normalizer module, making it part of the Normalizer type's impl. In
so doing we open the door for using additional state of the Normalizer
object from within this function.
